serializers.py
from rest_framework import serializers

from views import *
from models import *

from rest_framework_simplejwt.serializers import TokenObtainPairSerializer


class CustomTokenObtainPairSerializer(TokenObtainPairSerializer):
    @classmethod
    def get_token(cls, user):
        token = super().get_token(user)

        # Add custom claims
        token['username'] = user.username
        token['email'] = user.email

        return token


def create(validated_data):
    user = User.objects.create_user(**validated_data)
    return user


class SignUpSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True)

    class Meta:
        model = User
        fields = ('email', 'password', 'first_name', 'last_name')


class AdvertisementSerializer(serializers.ModelSerializer):
    class Meta:
        model = Advertisement
        fields = ['id', 'title', 'description', 'ad_creative', 'targeting_parameters', 'ad_placement']


class UserProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['user_id', 'user', 'bio', 'birthdate']


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = []
