requirements.txt
pip~=23.1.2
attrs~=23.1.0
wheel~=0.40.0
Pillow~=9.5.0
tornado~=6.3.2
Jinja2~=3.1.2
setuptools~=65.5.0
packaging~=23.1
pyparsing~=3.0.9
Django~=4.2.2
apiview~=1.3.25
djangorestframework~=3.14.0
Markdown~=3.4.3
future~=0.18.3
serializers~=0.2.4
rest_framework_simplejwt